fix: remove console log
