+++ src/index.tsx
@@ -2,7 +2,6 @@ import 'whatwg-fetch';
 
 import './app.css';
 
-import React from 'react';
 import ReactDOM from 'react-dom';
 import { HashRouter, Route } from 'react-router-dom';
 import { Provider } from 'react-redux';
@@ -14,7 +13,7 @@ import { StylesProvider } from '@material-ui/core/styles';
 import mainTheme from './themes/main-theme';
 import store from './store';
 import MetricsPoller from './metrics-poller';
-import App from './component/app';
+import App from './component/App';
 import ScrollToTop from './component/scroll-to-top';
 import { writeWarning } from './security-logger';
